requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
email-validator==2.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
Pillow
